More dashboard testing and tweaks

+++ database/migrations/2015_04_13_195050_create_snapshot_table.php
@@ -16,20 +16,23 @@ class CreateSnapshotTable extends Migration
      */
     public function up()
     {
-        \Schema::create(
-            'snapshot_t',
-            function ( Blueprint $table )
-            {
-                $table->bigIncrements( 'id' );
-                $table->string( 'snapshot_id_text', 128 );
-                $table->integer( 'user_id' )->index();
-                $table->integer( 'instance_id' )->index();
-                $table->string( 'url_text', 1024 );
-                $table->dateTime( 'expire_date' );
-                $table->dateTime( 'create_date' );
-                $table->timestamp( 'lmod_date' )->default( \DB::raw( 'CURRENT_TIMESTAMP' ) );
-            }
-        );
+        if ( !\Schema::hasTable( 'snapshot_t' ) )
+        {
+            \Schema::create(
+                'snapshot_t',
+                function ( Blueprint $table )
+                {
+                    $table->bigIncrements( 'id' );
+                    $table->string( 'snapshot_id_text', 128 );
+                    $table->integer( 'user_id' )->index();
+                    $table->integer( 'instance_id' )->index();
+                    $table->string( 'url_text', 1024 );
+                    $table->dateTime( 'expire_date' );
+                    $table->dateTime( 'create_date' );
+                    $table->timestamp( 'lmod_date' )->default( \DB::raw( 'CURRENT_TIMESTAMP' ) );
+                }
+            );
+        }
     }
 
     /**
@@ -39,7 +42,10 @@ class CreateSnapshotTable extends Migration
      */
     public function down()
     {
-        \Schema::drop( 'snapshot_t' );
+        if ( \Schema::hasTable( 'snapshot_t' ) )
+        {
+            \Schema::drop( 'snapshot_t' );
+        }
     }
 
 }
